project/project/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
